+++ blockchair/src/adapter.ts
@@ -1,0 +1,41 @@
+import { Requester, Validator, AdapterError } from '@chainlink/external-adapter'
+import { ExecuteWithConfig, ExecuteFactory, Config } from '@chainlink/types'
+import { makeConfig, DEFAULT_ENDPOINT } from './config'
+import { stats, balance } from './endpoint'
+
+const inputParams = {
+  endpoint: false,
+}
+
+// Export function to integrate with Chainlink node
+const execute: ExecuteWithConfig<Config> = async (request, config) => {
+  const validator = new Validator(request, inputParams)
+  if (validator.error) throw validator.error
+
+  Requester.logConfig(config)
+
+  const jobRunID = validator.validated.id
+  const endpoint = validator.validated.data.endpoint || DEFAULT_ENDPOINT
+
+  switch (endpoint) {
+    // might be moved to validator or config
+    case 'difficulty':
+    case 'height': {
+      return stats.execute(request, config)
+    }
+    case balance.Name: {
+      return balance.makeExecute(config)(request)
+    }
+    default: {
+      throw new AdapterError({
+        jobRunID,
+        message: `Endpoint ${endpoint} not supported.`,
+        statusCode: 400,
+      })
+    }
+  }
+}
+
+export const makeExecute: ExecuteFactory<Config> = (config) => {
+  return async (request) => execute(request, config || makeConfig())
+}