+++ blockchair/test/balance.test.ts
@@ -1,0 +1,10 @@
+import { makeExecute } from '../src/adapter'
+import { shouldBehaveLikeBalanceAdapter } from '@chainlink/adapter-test-helpers'
+
+shouldBehaveLikeBalanceAdapter(makeExecute(), [
+  'bitcoin_mainnet',
+  'dash_mainnet',
+  'doge_mainnet',
+  'litecoin_mainnet',
+  'bitcoin_cash_mainnet',
+])