+++ cfbenchmarks/src/endpoint/index.ts
@@ -1,0 +1,1 @@
+export * as values from './values'